fix: refactor account details
- moved transactions to AccountTransactions component;
- moved reviews to AccountReviews component;
- request for transactions send only by admin;
- request for reviews send only by admin and for caregiver.
